% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramAppWithExport.R
\name{histogramAppWithExport}
\alias{histogramAppWithExport}
\title{Generate a small example shiny app}
\usage{
histogramAppWithExport(x, outputDir)
}
\arguments{
\item{x}{A numeric vector to create a histogram from.}

\item{outputDir}{A character scalar pointing to a folder where output
files will be saved.}
}
\value{
A shiny app object
}
\description{
Generate a small example shiny app
}
