% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramApp.R
\name{histogramApp}
\alias{histogramApp}
\title{Generate a small example shiny app}
\usage{
histogramApp(x)
}
\arguments{
\item{x}{A numeric vector to create a histogram from.}
}
\value{
A shiny app object
}
\description{
Generate a small example shiny app
}
